flight search and filtering.cpp
So, This program is all about  flight reservation and filtering process for Ethiopian Airlines that checks available flights from Addis Ababa to various international destinations, with some flights restricted due to certain reasons. It stores flight details (flight number, origin, destination, and price) in a structure and filters out the restricted destinations. The program displays available flights or informs the user if no flights are available.